Add Office365 managament API connector
It is adapted from the one in SicConfAPI. The connector uses the
Microsoft API to fetch content and forward them to Sekoia.
The existing Office365 module has been split in two submodules to
separate the already-existing one called message_trace and the new one
called management_api.
As a difference with the old SicConfAPI version, no event caching is
done since we cache the last pull date and expect the API to not
duplicate events.